requirements.txt
celery==5.2.7
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
